LPD-69154 Fix class names that extends from BaseContentsSectionCMSTableFDSView
